combine settings components
